% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regenie_help_text.R
\name{get_regenie_help_text}
\alias{get_regenie_help_text}
\title{Get the text of the REGENIE help file}
\usage{
get_regenie_help_text(regenie_folder = get_regenie_folder())
}
\arguments{
\item{regenie_folder}{folder where REGENIE is installed}
}
\value{
the text of the REGENIE help file
}
\description{
Get the text of the REGENIE help file
}
