% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regenie_download_url.R
\name{get_regenie_source_download_url}
\alias{get_regenie_source_download_url}
\title{Get the REGENIE source code's download URL}
\usage{
get_regenie_source_download_url()
}
\description{
Get the REGENIE source code's download URL
}
