% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regenie_version.R
\name{get_regenie_version}
\alias{get_regenie_version}
\title{Get the version of REGENIE}
\usage{
get_regenie_version(regenie_folder = get_regenie_folder())
}
\arguments{
\item{regenie_folder}{folder where REGENIE is installed}
}
\value{
the version of REGENIE
}
\description{
Get the version of REGENIE
}
