% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_regenie.R
\name{install_regenie}
\alias{install_regenie}
\title{Determine if REGENIE is installed}
\usage{
install_regenie(
  url = "http://zzz.bwh.harvard.edu/regenie/dist/regenie-1.07-x86_64.zip",
  regenie_folder = get_regenie_folder()
)
}
\arguments{
\item{url}{REGENIE download URL}

\item{regenie_folder}{folder where REGENIE is installed}
}
\value{
Nothing
}
\description{
Determine if REGENIE is installed
}
