% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_regenie.R
\name{install_regenie}
\alias{install_regenie}
\title{Determine if REGENIE is installed}
\usage{
install_regenie(
  source_url = get_regenie_source_download_url(),
  exe_url = get_regenie_exe_download_url(),
  regenie_folder = get_regenie_folder()
)
}
\arguments{
\item{regenie_folder}{folder where REGENIE is installed}
}
\value{
Nothing
}
\description{
Determine if REGENIE is installed
}
