% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regenie.R
\name{run_regenie}
\alias{run_regenie}
\title{Run PLINK from arguments}
\usage{
run_regenie(args, regenie_folder = get_regenie_folder())
}
\arguments{
\item{args}{arguments for the PLINK executable}

\item{regenie_folder}{folder where PLINK is installed}
}
\value{
the text that PLINK returns
}
\description{
Run PLINK from arguments
}
