% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regenie.R
\name{run_regenie}
\alias{run_regenie}
\title{Run REGENIE from arguments}
\usage{
run_regenie(args, regenie_folder = get_regenie_folder())
}
\arguments{
\item{args}{arguments for the REGENIE executable}

\item{regenie_folder}{folder where REGENIE is installed}
}
\value{
the text that REGENIE returns
}
\description{
Run REGENIE from arguments
}
