% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_regenie.R
\name{uninstall_regenie}
\alias{uninstall_regenie}
\title{Uninstall REGENIE}
\usage{
uninstall_regenie(regenie_folder = get_regenie_folder())
}
\arguments{
\item{regenie_folder}{folder where REGENIE is installed}
}
\value{
Nothing
}
\description{
Uninstall REGENIE
}
