% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_regenie.R
\name{install_regenie}
\alias{install_regenie}
\title{Determine if REGENIE is installed}
\usage{
install_regenie(
  url = get_regenie_download_url(),
  regenie_folder = get_regenie_folder()
)
}
\arguments{
\item{url}{REGENIE download URL,
as can be obtained by \link{get_regenie_download_url}}

\item{regenie_folder}{folder where REGENIE is installed}
}
\value{
Nothing
}
\description{
Determine if REGENIE is installed
}
